backend/python/requirements.txt
transformers
spacy
sentence-transformers
weaviate-client<4.0.0
python-dotenv
google-genai
psycopg2-binary
pgvector

fastapi
uvicorn
requests
Pillow
jaclang==0.9.3